+++ packages/terra-content-container/CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## Unreleased
 
+* Changed
+  * Update wdio tests to remove deprecated Terra.it helpers
+
 ## 3.30.0 - (August 4, 2020)
 
 * Changed